Combined article and card components
